+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `HarpData` does not match csv if the file name is only `harpmessage.bin`. 
   See issue #93
 - Adapt `add_mouse` to new MCMS page layout
+- `flexiznam config --update` works even if the `secret_password` file does not exists
 
 ### Minor
 - add `compare_dictionaries_recursively` in `utils`